add - I split ModelGenerator into new and old so I can run side by side comparison.
